refactoring code to move filter code  within a component
