+++ pages/index.js
@@ -2,13 +2,12 @@ import useSWR from "swr";
 import styled, { css } from "styled-components";
 
 import PlantList from "@/components/PlantList";
-import PlantFilterForm from "@/components/PlantFilterForm";
 import { useState } from "react";
 import PlantCounter from "@/components/PlantCounter";
+import PlantFilter from "@/components/PlantFilter";
 
 export default function HomePage() {
   const { data, isLoading } = useSWR("/api/plants");
-  const [showFilter, setShowFilter] = useState(false);
   const [filteredPlants, setFilteredPlants] = useState();
 
   if (isLoading) {
@@ -18,6 +17,7 @@ export default function HomePage() {
   if (!data) {
     return <p>Failed to load plants!</p>;
   }
+
   if (data.length === 0) {
     return (
       <StyledMessage>
@@ -31,90 +31,16 @@ export default function HomePage() {
     );
   }
 
-  function handleFilterSubmit(event) {
-    event.preventDefault();
-
-    const formData = new FormData(event.target);
-
-    const lightNeed = formData.getAll("lightNeed");
-    const waterNeed = formData.getAll("waterNeed");
-
-    const filtered = data.filter((plant) => {
-      const matchesLight =
-        lightNeed.length === 0 || lightNeed.includes(plant.lightNeed);
-      const matchesWater =
-        waterNeed.length === 0 || waterNeed.includes(plant.waterNeed);
-
-      return matchesLight && matchesWater;
-    });
-    setFilteredPlants(filtered);
-  }
-
-  function handleClearSubmit(event) {
-    setFilteredPlants();
-  }
-
   const filteredPlantList = filteredPlants ?? data;
 
   return (
     <>
-      <FilterContainer $showFilter={showFilter}>
-        <ButtonContainer>
-          <FilterButton
-            $showFilter={showFilter}
-            onClick={() => {
-              setShowFilter(!showFilter);
-            }}
-          >
-            {!showFilter ? "Filter ⬇️" : "Filter ⬆️"}
-          </FilterButton>
-        </ButtonContainer>
-        {showFilter && (
-          <PlantFilterForm
-            onSubmit={handleFilterSubmit}
-            onClear={handleClearSubmit}
-          />
-        )}
-      </FilterContainer>
-
+      <PlantFilter data={data} setFilteredPlants={setFilteredPlants} />
       <PlantCounter length={filteredPlantList.length} />
-
       <PlantList plants={filteredPlantList} />
     </>
   );
 }
-
-const FilterContainer = styled.div`
-  ${(props) =>
-    props.$showFilter &&
-    css`
-      border: 1px solid black;
-      border-radius: 15px;
-      overflow: hidden;
-    `};
-`;
-
-const ButtonContainer = styled.div`
-  display: grid;
-  grid-template-columns: repeat(2, 1fr);
-  gap: 1rem;
-`;
-
-const FilterButton = styled.button`
-  width: 100%;
-  padding: 10px 20px;
-  background-color: lightgray;
-  border-radius: 15px;
-  font-size: medium;
-  ${(props) =>
-    props.$showFilter
-      ? css`
-          border: none;
-        `
-      : css`
-          border: 1px solid black;
-        `};
-`;
 
 const StyledMessage = styled.div`
   text-align: center;
